+++ package.json
@@ -6,13 +6,14 @@
   "scripts": {
     "lint": "eslint",
     "lint:fix": "eslint --fix",
-    "unit": "node --test test.js",
+    "unit": "node --test test/*.js",
     "test": "npm run lint && npm run unit",
-    "coverage": "nyc --reporter=lcov node --test test.js",
+    "coverage": "c8 --reporter=lcov node --test test.js",
+    "postcoverage": "c8 check-coverage --lines 100 --functions 100 --branches 100",
     "test:ci": "npm run lint && npm run coverage",
     "license-checker": "license-checker --production --onlyAllow='MIT;ISC;BSD-3-Clause;BSD-2-Clause;0BSD;Apache-2.0;Apache*'",
     "release": "read -p 'GITHUB_TOKEN: ' GITHUB_TOKEN && export GITHUB_TOKEN=$GITHUB_TOKEN && release-it --disable-metrics",
-    "mongodb": "docker run -d -p 27017:27017 mongo:latest mongod --setParameter ttlMonitorSleepSecs=2"
+    "mongodb": "docker run -d --name mongodb --rm -p 27017:27017 mongo:latest mongod --setParameter ttlMonitorSleepSecs=2"
   },
   "release-it": {
     "github": {
@@ -54,11 +55,11 @@
   "homepage": "https://github.com/moscajs/aedes-persistence-mongodb#readme",
   "devDependencies": {
     "@fastify/pre-commit": "^2.2.0",
-    "eslint": "^9.22.0",
+    "c8": "^10.1.3",
+    "eslint": "^9.24.0",
     "license-checker": "^25.0.1",
-    "mqemitter-mongodb": "^8.1.0",
+    "mqemitter-mongodb": "^9.0.0",
     "neostandard": "^0.12.1",
-    "nyc": "^17.1.0",
     "release-it": "^18.1.2"
   },
   "dependencies": {
